pyproject.toml
[project]
name = "photo-share-rest-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi[standard] (>=0.136.1,<0.137.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.0.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.27,<0.0.28)",
    "fastapi-limiter (>=0.2.0,<0.3.0)",
    "cloudinary (>=1.44.2,<2.0.0)",
    "redis (>=7.4.0,<8.0.0)",
    "fastapi-mail (>=1.6.4,<2.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.6.0,<5.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
]

[tool.black]
line-length = 70

[tool.ruff]
line-length = 70
